[change] Moves shared/ folder into components folder.
