+++ src/components/shared/icons/File.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { IconProps } from '../../types/icon';
+import { IconProps } from '../../../types/icon';
 
 export function File(props: IconProps) {
     const { color, size } = props;
