+++ src/components/shared/icons/GitHub.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { IconProps } from '../../types/icon';
+import { IconProps } from '../../../types/icon';
 
 export function GitHub(props: IconProps) {
     const { color, size } = props;
