+++ src/components/shared/icons/LinkedIn.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { IconProps } from '../../types/icon';
+import { IconProps } from '../../../types/icon';
 
 export function LinkedIn(props: IconProps) {
     const { color, size } = props;
