+++ src/components/shared/icons/Twitter.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { IconProps } from '../../types/icon';
+import { IconProps } from '../../../types/icon';
 
 export function Twitter(props: IconProps) {
     const { color, size } = props;
